+++ source/bxgenvtx/testing/test_tube_vg.cxx
@@ -1,4 +1,4 @@
-// -*- mode: c++ ; -*- 
+// -*- mode: c++ ; -*-
 // test_tube_vg.cxx
 
 #include <cstdlib>
@@ -17,8 +17,8 @@ int main (int argc_, char ** argv_)
   int error_code = EXIT_SUCCESS;
   try
     {
-      clog << "Test program for class 'tube_vg'!" << endl; 
-  
+      clog << "Test program for class 'tube_vg'!" << endl;
+
       bool debug = false;
 
       int iarg = 1;
@@ -28,21 +28,21 @@ int main (int argc_, char ** argv_)
 
           if (token[0] == '-')
             {
-               string option = token; 
-               if ((option == "-d") || (option == "--debug")) 
+               string option = token;
+               if ((option == "-d") || (option == "--debug"))
                  {
                    debug = true;
                  }
-               else 
-                 { 
-                    clog << "warning: ignoring option '" << option << "'!" << endl; 
+               else
+                 {
+                    clog << "warning: ignoring option '" << option << "'!" << endl;
                  }
             }
           else
             {
-              string argument = token; 
-               { 
-                clog << "warning: ignoring argument '" << argument << "'!" << endl; 
+              string argument = token;
+               {
+                clog << "warning: ignoring argument '" << argument << "'!" << endl;
               }
             }
           iarg++;
@@ -51,19 +51,19 @@ int main (int argc_, char ** argv_)
       std::string       rng_id = "mt19937";
       unsigned long int rng_seed = 12345;
       mygsl::rng random (rng_id, rng_seed);
-         
+
       geomtools::tube t (2.0, 3.0, 4.0);
 
       genvtx::tube_vg vg;
       vg.set_debug (debug);
 
       {
-        geomtools::vector_3d pos; 
+        geomtools::vector_3d pos;
         geomtools::rotation_3d rot;
         geomtools::gnuplot_draw::draw_tube (cout, pos, rot, t);
         cout << endl << endl;
       }
- 
+
       {
         vg.set_tube (t);
         vg.set_mode (genvtx::tube_vg::MODE_SURFACE);
@@ -78,8 +78,8 @@ int main (int argc_, char ** argv_)
         vg.initialize_simple ();
         vg.tree_dump (clog, "Tube vertex generator (inner side/bottom surface)");
         size_t nshoots = 5000;
-        geomtools::vector_3d vertex;     
-        for (int i = 0; i < nshoots; i++)
+        geomtools::vector_3d vertex;
+        for (int i = 0; i < (int) nshoots; i++)
           {
             vg.shoot_vertex (random, vertex);
             geomtools::gnuplot_draw::basic_draw_point (cout, vertex, true);
@@ -95,8 +95,8 @@ int main (int argc_, char ** argv_)
         vg.initialize_simple ();
         vg.tree_dump (clog, "Tube vertex generator (bulk)");
         size_t nshoots = 5000;
-        geomtools::vector_3d vertex;     
-        for (int i = 0; i < nshoots; i++)
+        geomtools::vector_3d vertex;
+        for (int i = 0; i < (int) nshoots; i++)
           {
             vg.shoot_vertex (random, vertex);
             geomtools::gnuplot_draw::basic_draw_point (cout, vertex, true);
@@ -107,12 +107,12 @@ int main (int argc_, char ** argv_)
     }
   catch (exception & x)
     {
-      cerr << "error: " << x.what () << endl; 
+      cerr << "error: " << x.what () << endl;
       error_code = EXIT_FAILURE;
     }
   catch (...)
     {
-      cerr << "error: " << "unexpected error!" << endl; 
+      cerr << "error: " << "unexpected error!" << endl;
       error_code = EXIT_FAILURE;
     }
   return (error_code);
